atributos referentes aos bolos e classe Util ppara métodos gerais
