--- backend/src/main/java/com/cakedeliver/cakedeliver/entities/Bolo.java
@@ -1,110 +1,0 @@
-package com.cakedeliver.cakedeliver.entities;
-
-import java.io.Serializable;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "tb_bolo")
-public class Bolo implements Serializable{
-
-	private static final long serialVersionUID = 1L;
-	
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY) //auto incremento no banco
-	
-	private Long id;
-	private String nome;
-	private Double preco;
-	private String descricao;
-	private String imagemUrl;
-	private Double tamanho;
-
-	public Bolo() {}
-
-	public Bolo(Long id, String nome, Double preco, String descricao, String imagemUrl, Double tamanho) {
-		super();
-		this.id = id;
-		this.nome = nome;
-		this.preco = preco;
-		this.descricao = descricao;
-		this.imagemUrl = imagemUrl;
-		this.tamanho = tamanho;
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getNome() {
-		return nome;
-	}
-
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
-
-	public Double getPreco() {
-		return preco;
-	}
-
-	public void setPreco(Double preco) {
-		this.preco = preco;
-	}
-
-	public String getDescricao() {
-		return descricao;
-	}
-
-	public void setDescricao(String descricao) {
-		this.descricao = descricao;
-	}
-
-	public String getImagemUrl() {
-		return imagemUrl;
-	}
-
-	public void setImagemUrl(String imagemUrl) {
-		this.imagemUrl = imagemUrl;
-	}
-
-	public Double getTamanho() {
-		return tamanho;
-	}
-
-	public void setTamanho(Double tamanho) {
-		this.tamanho = tamanho;
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((id == null) ? 0 : id.hashCode());
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		Bolo other = (Bolo) obj;
-		if (id == null) {
-			if (other.id != null)
-				return false;
-		} else if (!id.equals(other.id))
-			return false;
-		return true;
-	}
-}
