+++ backend/src/main/java/com/cakedeliver/cakedeliver/entities/Pedido.java
@@ -18,7 +18,7 @@ import com.cakedeliver.cakedeliver.enums.PedidoStatus;
 
 @Entity
 @Table(name = "tb_pedido")
-public class Pedido implements Serializable {
+public class Pedido extends Util implements Serializable {
 	
 	private static final long serialVersionUID = 1L;
 	
@@ -104,30 +104,5 @@ public class Pedido implements Serializable {
 
 	public Set<Bolo> getBolos() {
 		return bolos;
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((id == null) ? 0 : id.hashCode());
-		return result;
-	}
-	
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		Pedido other = (Pedido) obj;
-		if (id == null) {
-			if (other.id != null)
-				return false;
-		} else if (!id.equals(other.id))
-			return false;
-		return true;
 	}
 }
