--- backend/src/main/java/com/cakedeliver/cakedeliver/entities/Sabor.java
@@ -1,6 +1,0 @@
-package com.cakedeliver.cakedeliver.entities;
-
-public class Sabor {
-
-	private String nome;
-}
