--- backend/src/main/java/com/cakedeliver/cakedeliver/entities/Tamanho.java
@@ -1,7 +1,0 @@
-package com.cakedeliver.cakedeliver.entities;
-
-public class Tamanho {
-
-  private Integer tamanhoId;	
-	private Double tamanho;
-}
