+++ backend/src/main/java/com/cakedeliver/cakedeliver/entities/atores/Usuario.java
@@ -16,7 +16,7 @@ import com.cakedeliver.cakedeliver.enums.GeneroStatus;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
 @MappedSuperclass //para a heranças das classes que herdarao
-public abstract class Usuario implements Serializable {
+public abstract class Usuario extends Util implements Serializable {
 		
 	private static final long serialVersionUID = 1L;
 	
@@ -98,30 +98,5 @@ public abstract class Usuario implements Serializable {
   
   public Boolean isAtivo() {
 		return this.ativo;
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((id == null) ? 0 : id.hashCode());
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		Usuario other = (Usuario) obj;
-		if (id == null) {
-			if (other.id != null)
-				return false;
-		} else if (!id.equals(other.id))
-			return false;
-		return true;
 	}
 }
